new change for migration error

+++ database/migrations/2025_11_24_091922_create_role_permission_table.php
@@ -13,20 +13,22 @@ class CreateRolePermissionTable extends Migration
      */
     public function up()
     {
-        Schema::create('role_permission', function (Blueprint $table) {
-                        // $table->id();
+        if (! Schema::hasTable('role_permission')) {
+            Schema::create('role_permission', function (Blueprint $table) {
+                // $table->id();
 
-           $table->unsignedBigInteger('role_id');
-            $table->unsignedBigInteger('permission_id');
+                $table->unsignedBigInteger('role_id');
+                $table->unsignedBigInteger('permission_id');
 
-            // Composite primary key
-            $table->primary(['role_id', 'permission_id']);
+                // Composite primary key
+                $table->primary(['role_id', 'permission_id']);
 
-            // Foreign keys
-            $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-            $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
+                // Foreign keys
+                $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
+                $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
 
-        });
+            });
+        }
     }
 
     /**
